+++ packages/video_player/tizen/src/video_player.cc
@@ -57,8 +57,8 @@ void VideoPlayer::ReleaseMediaPacket(void *data) {
   player->SendRenderFinishedMessage();
 }
 
-FlutterDesktopGpuBuffer *VideoPlayer::ObtainGpuBuffer(size_t width,
-                                                      size_t height) {
+FlutterDesktopGpuSurfaceDescriptor *VideoPlayer::ObtainGpuSurface(
+    size_t width, size_t height) {
   std::lock_guard<std::mutex> lock(mutex_);
   if (!current_media_packet_) {
     LOG_ERROR("[VideoPlayer] current media packet not valid.");
@@ -77,12 +77,12 @@ FlutterDesktopGpuBuffer *VideoPlayer::ObtainGpuBuffer(size_t width,
     SendRenderFinishedMessage();
     return nullptr;
   }
-  flutter_desktop_gpu_buffer_->buffer = surface;
-  flutter_desktop_gpu_buffer_->width = width;
-  flutter_desktop_gpu_buffer_->height = height;
-  flutter_desktop_gpu_buffer_->release_context = this;
-  flutter_desktop_gpu_buffer_->release_callback = ReleaseMediaPacket;
-  return flutter_desktop_gpu_buffer_.get();
+  gpu_surface_->handle = surface;
+  gpu_surface_->width = width;
+  gpu_surface_->height = height;
+  gpu_surface_->release_context = this;
+  gpu_surface_->release_callback = ReleaseMediaPacket;
+  return gpu_surface_.get();
 }
 
 VideoPlayer::VideoPlayer(flutter::PluginRegistrar *plugin_registrar,
@@ -91,12 +91,13 @@ VideoPlayer::VideoPlayer(flutter::PluginRegistrar *plugin_registrar,
   texture_registrar_ = texture_registrar;
 
   texture_variant_ =
-      std::make_unique<flutter::TextureVariant>(flutter::GpuBufferTexture(
+      std::make_unique<flutter::TextureVariant>(flutter::GpuSurfaceTexture(
+          kFlutterDesktopGpuSurfaceTypeNone,
           [this](size_t width,
-                 size_t height) -> const FlutterDesktopGpuBuffer * {
-            return this->ObtainGpuBuffer(width, height);
+                 size_t height) -> const FlutterDesktopGpuSurfaceDescriptor * {
+            return this->ObtainGpuSurface(width, height);
           }));
-  flutter_desktop_gpu_buffer_ = std::make_unique<FlutterDesktopGpuBuffer>();
+  gpu_surface_ = std::make_unique<FlutterDesktopGpuSurfaceDescriptor>();
   texture_id_ = texture_registrar->RegisterTexture(texture_variant_.get());
 
   int ret = player_create(&player_);